# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treinamento API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/bruno/cart-items": {
            "get": {
                "tags": [
                    "Bruno Cart Items"
                ],
                "summary": "List all BrunoCartItems",
                "operationId": "80e21ab0b0f1bea776b0dcacd643e0b0",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bruno Cart Items"
                ],
                "summary": "Create a new BrunoCartItem",
                "operationId": "8070789116b754f332bbdaca1c005fee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "product_id": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/bruno/cart-items/{id}": {
            "get": {
                "tags": [
                    "Bruno Cart Items"
                ],
                "summary": "Get BrunoCartItem by ID",
                "operationId": "582b2230ef78efbc5abdb8a46ed07e22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bruno Cart Items"
                ],
                "summary": "Update a BrunoCartItem",
                "operationId": "0ba7c61f317032655eabcbbd1a37eb47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "product_id": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bruno Cart Items"
                ],
                "summary": "Delete a BrunoCartItem",
                "operationId": "0ca3addae90a885dc640d00acb540028",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/bruno/products": {
            "get": {
                "tags": [
                    "Bruno Products"
                ],
                "summary": "List all BrunoProducts",
                "operationId": "522da7bbb142934d43f6e07d6d8e78b6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bruno Products"
                ],
                "summary": "Create a new BrunoProduct",
                "operationId": "d99d59d9923437fc1832be2d44f9b9d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/bruno/products/{id}": {
            "get": {
                "tags": [
                    "Bruno Products"
                ],
                "summary": "Get BrunoProduct by ID",
                "operationId": "c232e9121f4950a99bbcd8b8deadc7b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bruno Products"
                ],
                "summary": "Update a BrunoProduct",
                "operationId": "017cab6cecd51d0f8aa8e85f6dbeb3c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bruno Products"
                ],
                "summary": "Delete a BrunoProduct",
                "operationId": "d9fa37cc8df8ad725203f01aad06ea40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}